+++ src/main/java/pl/example/repository/OpinionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.example.models.OpinionEntity;
 
 @Repository
-public interface OpinionRepository extends CrudRepository<OpinionEntity, Long> {
+public interface OpinionRepository extends CrudRepository<OpinionEntity, Integer> {
 }
